Move shared runtime parameters to runtime common package
