+++ runtime/common/Cargo.toml
@@ -9,6 +9,10 @@ repository = "https://github.com/xx-labs/xxchain/"
 [dependencies]
 
 # third-party dependencies
+codec = { package = "parity-scale-codec", version = "2.2.0", default-features = false, features = [
+	"derive",
+	"max-encoded-len",
+] }
 static_assertions = "1.1.0"
 smallvec = "1.7.0"
 
@@ -16,6 +20,9 @@ smallvec = "1.7.0"
 sp-std = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
 sp-runtime = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
 node-primitives = { version = "0.2.0", default-features = false, path = "../../primitives" }
+sp-staking = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
+sp-npos-elections = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
+sp-io = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
 
 # frame dependencies
 frame-support = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
@@ -26,17 +33,25 @@ pallet-authorship = { git = "https://github.com/xx-labs/substrate", branch = "xx
 pallet-balances = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
 pallet-transaction-payment = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
 pallet-treasury = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
+pallet-staking = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
+pallet-election-provider-multi-phase = { git = "https://github.com/xx-labs/substrate", branch = "xx-network", default-features = false }
 
 [features]
 default = ["std"]
 std = [
+	"codec/std",
 	"sp-std/std",
 	"pallet-authorship/std",
 	"pallet-balances/std",
 	"node-primitives/std",
+	"sp-staking/std",
+	"sp-npos-elections/std",
+	"sp-io/std",
 	"sp-runtime/std",
 	"frame-support/std",
 	"frame-system/std",
 	"pallet-transaction-payment/std",
 	"pallet-treasury/std",
+	"pallet-staking/std",
+	"pallet-election-provider-multi-phase/std",
 ]
